+++ src/test/kotlin/io/kjson/uri/URITemplateTest.kt
@@ -105,8 +105,8 @@ class URITemplateTest {
                     expect("(prefix)") { text }
                 }
                 with(this[1]) {
-                    assertIs<ExpressionElement>(this)
-                    expect("var") { name }
+                    assertIs<VariableElement>(this)
+                    expect(listOf("var")) { names }
                 }
                 with(this[2]) {
                     assertIs<TextElement>(this)
@@ -160,7 +160,7 @@ class URITemplateTest {
 
     @Test fun `should throw exception on empty expression`() {
         assertFailsWith<URITemplateException> { URITemplate.parse("(prefix){}(suffix)") }.let {
-            expect("Expression is empty at offset 9") { it.message }
+            expect("Variable name is empty at offset 9") { it.message }
         }
     }
 
@@ -188,8 +188,8 @@ class URITemplateTest {
                     expect("(prefix)") { text }
                 }
                 with(this[1]) {
-                    assertIs<ExpressionElement>(this)
-                    expect("var") { name }
+                    assertIs<VariableElement>(this)
+                    expect(listOf("var")) { names }
                 }
                 with(this[2]) {
                     assertIs<TextElement>(this)
@@ -212,13 +212,184 @@ class URITemplateTest {
         assertNull(uriTemplate["var"])
     }
 
+    @Test fun `should create template with reserved variable`() {
+        val uriTemplate = URITemplate.parse("(prefix){+var}(suffix)")
+        with(uriTemplate) {
+            with(elements) {
+                expect(3) { size }
+                with(this[0]) {
+                    assertIs<TextElement>(this)
+                    expect("(prefix)") { text }
+                }
+                with(this[1]) {
+                    assertIs<ReservedElement>(this)
+                    expect(listOf("var")) { names }
+                }
+                with(this[2]) {
+                    assertIs<TextElement>(this)
+                    expect("(suffix)") { text }
+                }
+            }
+            with(variables) {
+                expect(1) { size }
+                expect(Variable("var", null)) { this[0] }
+            }
+        }
+        assertTrue("var" in uriTemplate)
+        assertFalse("xxx" in uriTemplate)
+    }
+
+    @Test fun `should expand template with reserved variable`() {
+        val uriTemplate = URITemplate.parse("(prefix){+var}(suffix)")
+        uriTemplate["var"] = "/abc/def"
+        expect("(prefix)/abc/def(suffix)") { uriTemplate.toString() }
+        expect("/abc/def") { uriTemplate["var"] }
+    }
+
+    @Test fun `should create template with fragment variable`() {
+        val uriTemplate = URITemplate.parse("(prefix){#var}(suffix)")
+        with(uriTemplate) {
+            with(elements) {
+                expect(3) { size }
+                with(this[0]) {
+                    assertIs<TextElement>(this)
+                    expect("(prefix)") { text }
+                }
+                with(this[1]) {
+                    assertIs<FragmentElement>(this)
+                    expect(listOf("var")) { names }
+                }
+                with(this[2]) {
+                    assertIs<TextElement>(this)
+                    expect("(suffix)") { text }
+                }
+            }
+            with(variables) {
+                expect(1) { size }
+                expect(Variable("var", null)) { this[0] }
+            }
+        }
+        assertTrue("var" in uriTemplate)
+        assertFalse("xxx" in uriTemplate)
+    }
+
+    @Test fun `should expand template with fragment variable`() {
+        val uriTemplate = URITemplate.parse("(prefix){#var}(suffix)")
+        uriTemplate["var"] = "abc$"
+        expect("(prefix)#abc$(suffix)") { uriTemplate.toString() }
+        expect("abc$") { uriTemplate["var"] }
+    }
+
     @Test fun `should perform substitutions listed in specification for Level 1`() {
-        val uriTemplate1 = URITemplate.parse("{var}")
-        uriTemplate1["var"] = "value"
+        val uriTemplate1 = URITemplate.parse("{var}").apply {
+            this["var"] = "value"
+        }
         expect("value") { uriTemplate1.toString() }
-        val uriTemplate2 = URITemplate.parse("{hello}")
-        uriTemplate2["hello"] = "Hello World!"
+        val uriTemplate2 = URITemplate.parse("{hello}").apply {
+            this["hello"] = "Hello World!"
+        }
         expect("Hello%20World%21") { uriTemplate2.toString() }
+    }
+
+    @Test fun `should perform substitutions listed in specification for Level 2`() {
+        val uriTemplate1 = URITemplate.parse("{+var}").apply {
+            this["var"] = "value"
+        }
+        expect("value") { uriTemplate1.toString() }
+        val uriTemplate2 = URITemplate.parse("{+hello}").apply {
+            this["hello"] = "Hello World!"
+        }
+        expect("Hello%20World!") { uriTemplate2.toString() }
+        val uriTemplate3 = URITemplate.parse("{+path}/here").apply {
+            this["path"] = "/foo/bar"
+        }
+        expect("/foo/bar/here") { uriTemplate3.toString() }
+    }
+
+    @Test fun `should perform substitutions listed in specification for Level 3`() {
+        val uriTemplate1 = URITemplate.parse("map?{x,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+        }
+        expect("map?1024,768") { uriTemplate1.toString() }
+        val uriTemplate2 = URITemplate.parse("{x,hello,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["hello"] = "Hello World!"
+        }
+        expect("1024,Hello%20World%21,768") { uriTemplate2.toString() }
+        val uriTemplate3 = URITemplate.parse("{+x,hello,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["hello"] = "Hello World!"
+        }
+        expect("1024,Hello%20World!,768") { uriTemplate3.toString() }
+        val uriTemplate4 = URITemplate.parse("{+path,x}/here").apply {
+            this["x"] = "1024"
+            this["path"] = "/foo/bar"
+        }
+        expect("/foo/bar,1024/here") { uriTemplate4.toString() }
+        val uriTemplate5 = URITemplate.parse("{#x,hello,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["hello"] = "Hello World!"
+        }
+        expect("#1024,Hello%20World!,768") { uriTemplate5.toString() }
+        val uriTemplate6 = URITemplate.parse("{#path,x}/here").apply {
+            this["x"] = "1024"
+            this["path"] = "/foo/bar"
+        }
+        expect("#/foo/bar,1024/here") { uriTemplate6.toString() }
+        val uriTemplate7 = URITemplate.parse("X{.var}").apply {
+            this["var"] = "value"
+        }
+        expect("X.value") { uriTemplate7.toString() }
+        val uriTemplate8 = URITemplate.parse("X{.x,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+        }
+        expect("X.1024.768") { uriTemplate8.toString() }
+        val uriTemplate9 = URITemplate.parse("{/var}").apply {
+            this["var"] = "value"
+        }
+        expect("/value") { uriTemplate9.toString() }
+        val uriTemplate10 = URITemplate.parse("{/var,x}/here").apply {
+            this["var"] = "value"
+            this["x"] = "1024"
+        }
+        expect("/value/1024/here") { uriTemplate10.toString() }
+        val uriTemplate11 = URITemplate.parse("{;x,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+        }
+        expect(";x=1024;y=768") { uriTemplate11.toString() }
+        val uriTemplate12 = URITemplate.parse("{;x,y,empty}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["empty"] = ""
+        }
+        expect(";x=1024;y=768;empty") { uriTemplate12.toString() }
+        val uriTemplate13 = URITemplate.parse("{?x,y}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+        }
+        expect("?x=1024&y=768") { uriTemplate13.toString() }
+        val uriTemplate14 = URITemplate.parse("{?x,y,empty}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["empty"] = ""
+        }
+        expect("?x=1024&y=768&empty=") { uriTemplate14.toString() }
+        val uriTemplate15 = URITemplate.parse("?fixed=yes{&x}").apply {
+            this["x"] = "1024"
+        }
+        expect("?fixed=yes&x=1024") { uriTemplate15.toString() }
+        val uriTemplate16 = URITemplate.parse("{&x,y,empty}").apply {
+            this["x"] = "1024"
+            this["y"] = "768"
+            this["empty"] = ""
+        }
+        expect("&x=1024&y=768&empty=") { uriTemplate16.toString() }
     }
 
 }