+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.0] - 2024-10-14
+### Changed
+- `ExpressionElement`, `URITemplate`: added support for specification level 2
+
 ## [1.0] - 2024-10-14
 ### Added
-- all files: library complete to specification level 1
+- multiple files: library complete to specification level 1
 
 ## [0.1] - 2024-10-12
 ### Added